requirements.txt
Flask==1.1.1
google-cloud-datastore==1.11.0
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'